TensorInference

Documentation for TensorInference.

TensorInference.Factor

TensorInference.MMAPModel

TensorInference.MMAPModel

TensorInference.MMAPModel

TensorInference.RescaledArray

TensorInference.Samples

TensorInference.TensorNetworkModel

TensorInference.TensorNetworkModel

TensorInference.TensorNetworkModel

TensorInference.TensorNetworkModel

TensorInference.UAIInstance

TensorInference.backward_sampling!

TensorInference.backward_tropical

TensorInference.get_cards

TensorInference.get_cards

TensorInference.get_vars

TensorInference.get_vars

TensorInference.log_probability

TensorInference.marginals

TensorInference.maximum_logp

TensorInference.most_probable_config

TensorInference.parse_mar_solution_file

TensorInference.probability

TensorInference.read_evidence_file

TensorInference.read_instance

TensorInference.read_model_file

TensorInference.read_query_file

TensorInference.read_td_file

TensorInference.rescale_array

TensorInference.sample

TensorInference.set_evidence!

TensorInference.Factor — Type

struct Factor{T, N}

Fields

vars

vals

Encodes a discrete function over the set of variables vars that maps each instantiation of vars into a nonnegative number in vals.

TensorInference.MMAPModel — Type

struct MMAPModel{LT, AT<:AbstractArray}

Computing the most likely assignment to the query variables, Xₘ ⊆ X after marginalizing out the remaining variables Xₛ = X \ Xₘ.

\[{\rm MMAP}(X_i|E=e) = \arg \max_{X_M} \sum_{X_S} \prod_{F} f(x_M, x_S, e)\]

Fields

vars is the remaining (or not marginalized) degree of freedoms in the tensor network.

code is the tropical tensor network contraction pattern.

tensors is the tensors fed into the tensor network.

clusters is the clusters, each element of this cluster is a TensorNetworkModel instance for marginalizing certain variables.

fixedvertices is a dictionary to specifiy degree of freedoms fixed to certain values, which should not have overlap with the marginalized variables.

TensorInference.MMAPModel — Method

MMAPModel(instance::TensorInference.UAIInstance)

TensorInference.MMAPModel — Method

MMAPModel(
    vars::AbstractArray{LT, 1},
    factors::Array{<:TensorInference.Factor{T}, 1}
)

TensorInference.RescaledArray — Type

struct RescaledArray{T, N, AT<:AbstractArray{T, N}} <: AbstractArray{T, N}

RescaledArray(α, T) -> RescaledArray

An array data type with a log-prefactor, and a l∞-normalized storage, i.e. the maximum element in a tensor is 1. This tensor type can avoid the potential underflow/overflow of numbers in a tensor network. The constructor RescaledArray(α, T) creates a rescaled array that equal to exp(α) * T.

TensorInference.Samples — Type

struct Samples{L}

Fields

samples::Matrix{Int64}

labels::Vector

setmask::BitVector

The sampled configurations are stored in samples, which is a vector of vector. labels is a vector of variable names for labeling configurations. The setmask is an boolean indicator to denote whether the sampling process of a variable is complete.

TensorInference.TensorNetworkModel — Type

struct TensorNetworkModel{LT, ET, MT<:AbstractArray}

Probabilistic modeling with a tensor network.

Fields

vars is the degree of freedoms in the tensor network.

code is the tensor network contraction pattern.

tensors is the tensors fed into the tensor network.

fixedvertices is a dictionary to specifiy degree of freedoms fixed to certain values.

TensorInference.TensorNetworkModel — Method

TensorNetworkModel(
    instance::TensorInference.UAIInstance
) -> TensorNetworkModel{Int64}

TensorInference.TensorNetworkModel — Method

TensorNetworkModel(
    vars::AbstractArray{LT, 1},
    rawcode::OMEinsum.EinCode,
    tensors::Vector{<:AbstractArray}
) -> TensorNetworkModel

TensorInference.TensorNetworkModel — Method

TensorNetworkModel(
    vars::AbstractArray{LT, 1},
    cards::AbstractVector{Int64},
    factors::Array{<:TensorInference.Factor{T}, 1}
) -> TensorNetworkModel

TensorInference.UAIInstance — Type

struct UAIInstance{ET, FT<:(TensorInference.Factor{ET})}

Fields

nvars is the number of variables,

nclique is the number of cliques,

cards is a vector of cardinalities for variables,

factors is a vector of factors,

obsvars is a vector of observed variables,

obsvals is a vector of observed values,

queryvars is a vector of query variables,

reference_solution is a vector with the reference solution.

TensorInference.backward_sampling! — Method

backward_sampling!(
    ixs,
    xs::Tuple,
    iy,
    y,
    samples::TensorInference.Samples,
    size_dict
) -> TensorInference.Samples

The backward process for sampling configurations.

ixs and xs are labels and tensor data for input tensors,

iy and y are labels and tensor data for the output tensor,

samples is the samples generated for eliminated variables,

size_dict is a key-value map from tensor label to dimension size.

TensorInference.backward_tropical — Method

backward_tropical(
    ixs,
    xs::Tuple,
    iy,
    y,
    ymask,
    size_dict
) -> Vector{Any}

The backward rule for tropical einsum.

ixs and xs are labels and tensor data for input tensors,

iy and y are labels and tensor data for the output tensor,

ymask is the boolean mask for gradients,

size_dict is a key-value map from tensor label to dimension size.

TensorInference.get_cards — Method

get_cards(mmap::MMAPModel) -> Vector

TensorInference.get_cards — Method

get_cards(tn::TensorNetworkModel) -> Vector

Get the cardinalities of variables in this tensor network.

TensorInference.get_vars — Method

get_vars(mmap::MMAPModel) -> Vector

TensorInference.get_vars — Method

get_vars(tn::TensorNetworkModel) -> Vector

Get the variables in this tensor network, they are also known as legs, labels, or degree of freedoms.

TensorInference.log_probability — Method

log_probability(
    tn::TensorNetworkModel,
    config::Union{Dict, AbstractVector}
) -> Real

Evaluate the log probability of config.

TensorInference.marginals — Method

marginals(tn::TensorNetworkModel) -> Vector

Returns the marginal probability distribution of variables. One can use get_vars(tn) to get the full list of variables in this tensor network.

TensorInference.maximum_logp — Method

maximum_logp(
    tn::TensorNetworkModel
) -> AbstractArray{<:Real}

Returns an output array containing largest log-probabilities.

TensorInference.most_probable_config — Method

most_probable_config(
    tn::TensorNetworkModel
) -> Tuple{Real, Vector}

Returns the largest log-probability and the most probable configuration.

TensorInference.parse_mar_solution_file — Method

parse_mar_solution_file(
    rawlines::Vector{String}
) -> Vector{Vector{Float64}}

Parse the solution marginals of all variables from the UAI MAR solution file. The order of the variables is the same as in the model definition.

The UAI file formats are defined in: https://uaicompetition.github.io/uci-2022/file-formats/

TensorInference.probability — Method

probability(tn::TensorNetworkModel) -> AbstractArray

Contract the tensor network and return a probability array with its rank specified in the contraction code tn.code. The returned array may not be l1-normalized even if the total probability is l1-normalized, because the evidence tn.fixedvertices may not be empty.

TensorInference.read_evidence_file — Method

read_evidence_file(
    evidence_filepath::AbstractString
) -> Tuple{Vector{Int64}, Vector{Int64}}

Return the observed variables and values in evidence_filepath. If the passed file path is an empty string, return empty vectors.

The UAI file formats are defined in: https://uaicompetition.github.io/uci-2022/file-formats/

TensorInference.read_instance — Method

read_instance(
    model_filepath::AbstractString
) -> TensorInference.UAIInstance

Read a UAI problem instance from a file.

TensorInference.read_model_file — Method

read_model_file(
    model_filepath
) -> Tuple{Int64, Vector{Int64}, Int64, Any}

Parse the problem instance found in model_filepath defined in the UAI model format.

The UAI file formats are defined in: https://uaicompetition.github.io/uci-2022/file-formats/

TensorInference.read_query_file — Method

read_query_file(
    query_filepath::AbstractString
) -> Vector{Int64}

Return the query variables in query_filepath. If the passed file path is an empty string, return an empty vector.

The UAI file formats are defined in: https://uaicompetition.github.io/uci-2022/file-formats/

TensorInference.read_td_file — Method

read_td_file(
    td_filepath::AbstractString
) -> Tuple{Int64, Int64, Int64, Vector{Vector{Int64}}, Vector{Vector{Int64}}}

Parse a tree decomposition instance described the PACE format.

The PACE file format is defined in: https://pacechallenge.org/2017/treewidth/

TensorInference.rescale_array — Method

Returns a rescaled array that equivalent to the input tensor.

TensorInference.sample — Method

sample(tn::TensorNetworkModel, n::Int64) -> Matrix{Int64}

Generate samples from a tensor network based probabilistic model. Returns a vector of vector, each element being a configurations defined on get_vars(tn).

Arguments

tn is the tensor network model.

n is the number of samples to be returned.

TensorInference.set_evidence! — Method

set_evidence!(
    uai::TensorInference.UAIInstance,
    pairs::Pair{Int64}...
) -> TensorInference.UAIInstance

Set the evidence of an UAI instance.