TensorInference

Documentation for TensorInference.

TensorInference.Factor

TensorInference.MMAPModeling

TensorInference.MMAPModeling

TensorInference.TensorNetworkModeling

TensorInference.TensorNetworkModeling

TensorInference.TensorNetworkModeling

TensorInference.backward_tropical

TensorInference.get_vars

TensorInference.get_vars

TensorInference.marginals

TensorInference.maximum_logp

TensorInference.maximum_logp

TensorInference.most_probable_config

TensorInference.most_probable_config

TensorInference.probability

TensorInference.probability

TensorInference.probability

TensorInference.read_td_file

TensorInference.read_uai_evid_file

TensorInference.read_uai_file

TensorInference.read_uai_mar_file

TensorInference.Factor — Type

struct Factor{T, N}

Fields

vars

vals

Encodes a discrete function over the set of variables vars that maps each instantiation of vars into a nonnegative number in vals.

TensorInference.MMAPModeling — Type

struct MMAPModeling{LT, AT<:AbstractArray}

Computing the most likely assignment to the query variables, Xₘ ⊆ X after marginalizing out the remaining variables Xₛ = X \ Xₘ.

\[{\rm MMAP}(X_i|E=e) = \arg \max_{X_M} \sum_{X_S} \prod_{F} f(x_M, x_S, e)\]

Fields

vars is the remaining (or not marginalized) degree of freedoms in the tensor network.

code is the tropical tensor network contraction pattern.

tensors is the tensors fed into the tensor network.

models is the clusters, each element of this cluster is a TensorNetworkModeling instance for marginalizing certain variables.

fixedvertices is a dictionary to specifiy degree of freedoms fixed to certain values, which should not have overlap with the marginalized variables.

TensorInference.MMAPModeling — Method

MMAPModeling(
    vars::AbstractArray{LT, 1},
    factors::Array{<:TensorInference.Factor{T}, 1};
    marginalizedvertices,
    openvertices,
    fixedvertices,
    optimizer,
    simplifier,
    marginalize_optimizer,
    marginalize_simplifier
)

TensorInference.TensorNetworkModeling — Type

struct TensorNetworkModeling{LT, ET, MT<:AbstractArray}

Probabilistic modeling with a tensor network.

Fields

vars is the degree of freedoms in the tensor network.

code is the tensor network contraction pattern.

tensors is the tensors fed into the tensor network.

fixedvertices is a dictionary to specifiy degree of freedoms fixed to certain values.

TensorInference.TensorNetworkModeling — Method

TensorNetworkModeling(
    vars::AbstractArray{LT, 1},
    rawcode::OMEinsum.EinCode,
    tensors::Vector{<:AbstractArray};
    fixedvertices,
    optimizer,
    simplifier
) -> TensorNetworkModeling

TensorInference.TensorNetworkModeling — Method

TensorInference.backward_tropical — Method

backward_tropical(
    ixs,
    xs::Tuple,
    iy,
    y,
    ymask,
    size_dict
) -> Vector{Any}

The backward rule for tropical einsum.

ixs and xs are labels and tensor data for input tensors,

iy and y are labels and tensor data for the output tensor,

ymask is the boolean mask for gradients,

size_dict is a key-value map from tensor label to dimension size.

TensorInference.get_vars — Method

get_vars(mmap::MMAPModeling) -> Vector

TensorInference.get_vars — Method

get_vars(tn::TensorNetworkModeling) -> Vector

Get the variables in this tensor network, they is also known as legs, labels, or degree of freedoms.

TensorInference.marginals — Method

marginals(tn::TensorNetworkModeling) -> Vector

Returns the marginal probability distribution of variables. One can use get_vars(tn) to get the full list of variables in this tensor network.

TensorInference.maximum_logp — Method

maximum_logp(
    mmap::MMAPModeling
) -> AbstractArray{<:TropicalNumbers.Tropical}

TensorInference.maximum_logp — Method

maximum_logp(
    tn::TensorNetworkModeling
) -> AbstractArray{<:TropicalNumbers.Tropical}

Returns an output array containing largest log-probabilities.

TensorInference.most_probable_config — Method

most_probable_config(
    mmap::MMAPModeling
) -> Tuple{TropicalNumbers.Tropical, Vector}

TensorInference.most_probable_config — Method

most_probable_config(
    tn::TensorNetworkModeling
) -> Tuple{TropicalNumbers.Tropical, Vector}

Returns the largest log-probability and the most probable configuration.

TensorInference.probability — Method

probability(mmap::MMAPModeling, config) -> Real

TensorInference.probability — Method

probability(tn::TensorNetworkModeling, config) -> Real

Evaluate the probability of config.

TensorInference.probability — Method

probability(tn::TensorNetworkModeling) -> AbstractArray

Contract the tensor network and return a probability array with its rank specified in the contraction code tn.code. The returned array may not be l1-normalized even if the total probability is l1-normalized, because the evidence tn.fixedvertices may not be empty.

TensorInference.read_td_file — Method

read_td_file(
    td_filepath::AbstractString
) -> Tuple{Int64, Int64, Int64, Vector{Vector{Int64}}, Vector{Vector{Int64}}}

Parse a tree decomposition instance described the PACE format.

The PACE file format is defined in: https://pacechallenge.org/2017/treewidth/

TensorInference.read_uai_evid_file — Method

read_uai_evid_file(
    uai_evid_filepath::AbstractString
) -> Tuple{Vector{Int64}, Vector{Int64}}

Return the observed variables and values in uai_evid_filepath. If the passed file path is an empty string, return empty vectors.

The UAI file formats are defined in: https://personal.utdallas.edu/~vibhav.gogate/uai16-evaluation/uaiformat.html

TensorInference.read_uai_file — Method

read_uai_file(
    uai_filepath;
    factor_eltype
) -> Tuple{Int64, Vector{Int64}, Int64, Any}

Parse the problem instance found in uai_filepath defined in the UAI model format.

The UAI file formats are defined in: https://personal.utdallas.edu/~vibhav.gogate/uai16-evaluation/uaiformat.html

TensorInference.read_uai_mar_file — Method

read_uai_mar_file(
    uai_mar_filepath::AbstractString;
    factor_eltype
) -> Vector{Vector{Float64}}

Return the marginals of all variables. The order of the variables is the same as in the model

The UAI file formats are defined in: https://personal.utdallas.edu/~vibhav.gogate/uai16-evaluation/uaiformat.html