TensorInference

Documentation for TensorInference.

TensorInference.Factor

TensorInference.MMAPModel

TensorInference.MMAPModel

TensorInference.MMAPModel

TensorInference.RescaledArray

TensorInference.TensorNetworkModel

TensorInference.TensorNetworkModel

TensorInference.TensorNetworkModel

TensorInference.TensorNetworkModel

TensorInference.UAIInstance

TensorInference.backward_tropical

TensorInference.get_cards

TensorInference.get_cards

TensorInference.get_vars

TensorInference.get_vars

TensorInference.log_probability

TensorInference.marginals

TensorInference.maximum_logp

TensorInference.most_probable_config

TensorInference.probability

TensorInference.read_td_file

TensorInference.read_uai_evid_file

TensorInference.read_uai_file

TensorInference.read_uai_mar_file

TensorInference.read_uai_problem

TensorInference.rescale_array

TensorInference.set_evidence!

TensorInference.uai_problem_from_file

TensorInference.Factor — Type

struct Factor{T, N}

Fields

vars

vals

Encodes a discrete function over the set of variables vars that maps each instantiation of vars into a nonnegative number in vals.

TensorInference.MMAPModel — Type

struct MMAPModel{LT, AT<:AbstractArray}

Computing the most likely assignment to the query variables, Xₘ ⊆ X after marginalizing out the remaining variables Xₛ = X \ Xₘ.

\[{\rm MMAP}(X_i|E=e) = \arg \max_{X_M} \sum_{X_S} \prod_{F} f(x_M, x_S, e)\]

Fields

vars is the remaining (or not marginalized) degree of freedoms in the tensor network.

code is the tropical tensor network contraction pattern.

tensors is the tensors fed into the tensor network.

clusters is the clusters, each element of this cluster is a TensorNetworkModel instance for marginalizing certain variables.

fixedvertices is a dictionary to specifiy degree of freedoms fixed to certain values, which should not have overlap with the marginalized variables.

TensorInference.MMAPModel — Method

MMAPModel(instance::TensorInference.UAIInstance)

TensorInference.MMAPModel — Method

MMAPModel(
    vars::AbstractArray{LT, 1},
    factors::Array{<:TensorInference.Factor{T}, 1}
)

TensorInference.RescaledArray — Type

struct RescaledArray{T, N, AT<:AbstractArray{T, N}} <: AbstractArray{T, N}

RescaledArray(α, T) -> RescaledArray

An array data type with a log-prefactor, and a l∞-normalized storage, i.e. the maximum element in a tensor is 1. This tensor type can avoid the potential underflow/overflow of numbers in a tensor network. The constructor RescaledArray(α, T) creates a rescaled array that equal to exp(α) * T.

TensorInference.TensorNetworkModel — Type

struct TensorNetworkModel{LT, ET, MT<:AbstractArray}

Probabilistic modeling with a tensor network.

Fields

vars is the degree of freedoms in the tensor network.

code is the tensor network contraction pattern.

tensors is the tensors fed into the tensor network.

fixedvertices is a dictionary to specifiy degree of freedoms fixed to certain values.

TensorInference.TensorNetworkModel — Method

TensorNetworkModel(
    instance::TensorInference.UAIInstance
) -> TensorNetworkModel{Int64}

TensorInference.TensorNetworkModel — Method

TensorNetworkModel(
    vars::AbstractArray{LT, 1},
    rawcode::OMEinsum.EinCode,
    tensors::Vector{<:AbstractArray}
) -> TensorNetworkModel

TensorInference.TensorNetworkModel — Method

TensorNetworkModel(
    vars::AbstractArray{LT, 1},
    cards::AbstractVector{Int64},
    factors::Array{<:TensorInference.Factor{T}, 1}
) -> TensorNetworkModel

TensorInference.UAIInstance — Type

struct UAIInstance{ET, FT<:(TensorInference.Factor{ET})}

Fields

nvars is the number of variables,

nclique is the number of cliques,

cards is a vector of cardinalities for variables,

factors is a vector of factors,

obsvars is a vector of observed variables,

obsvals is a vector of observed values,

reference_marginals is a vector of marginal probabilities.

TensorInference.backward_tropical — Method

backward_tropical(
    ixs,
    xs::Tuple,
    iy,
    y,
    ymask,
    size_dict
) -> Vector{Any}

The backward rule for tropical einsum.

ixs and xs are labels and tensor data for input tensors,

iy and y are labels and tensor data for the output tensor,

ymask is the boolean mask for gradients,

size_dict is a key-value map from tensor label to dimension size.

TensorInference.get_cards — Method

get_cards(mmap::MMAPModel) -> Vector

TensorInference.get_cards — Method

get_cards(tn::TensorNetworkModel) -> Vector

Get the cardinalities of variables in this tensor network.

TensorInference.get_vars — Method

get_vars(mmap::MMAPModel) -> Vector

TensorInference.get_vars — Method

get_vars(tn::TensorNetworkModel) -> Vector

Get the variables in this tensor network, they are also known as legs, labels, or degree of freedoms.

TensorInference.log_probability — Method

log_probability(
    tn::TensorNetworkModel,
    config::Union{Dict, AbstractVector}
) -> Real

Evaluate the log probability of config.

TensorInference.marginals — Method

marginals(tn::TensorNetworkModel) -> Vector

Returns the marginal probability distribution of variables. One can use get_vars(tn) to get the full list of variables in this tensor network.

TensorInference.maximum_logp — Method

maximum_logp(
    tn::TensorNetworkModel
) -> AbstractArray{<:Real}

Returns an output array containing largest log-probabilities.

TensorInference.most_probable_config — Method

most_probable_config(
    tn::TensorNetworkModel
) -> Tuple{Real, Vector}

Returns the largest log-probability and the most probable configuration.

TensorInference.probability — Method

probability(tn::TensorNetworkModel) -> AbstractArray

Contract the tensor network and return a probability array with its rank specified in the contraction code tn.code. The returned array may not be l1-normalized even if the total probability is l1-normalized, because the evidence tn.fixedvertices may not be empty.

TensorInference.read_td_file — Method

read_td_file(
    td_filepath::AbstractString
) -> Tuple{Int64, Int64, Int64, Vector{Vector{Int64}}, Vector{Vector{Int64}}}

Parse a tree decomposition instance described the PACE format.

The PACE file format is defined in: https://pacechallenge.org/2017/treewidth/

TensorInference.read_uai_evid_file — Method

read_uai_evid_file(
    uai_evid_filepath::AbstractString
) -> Tuple{Vector{Int64}, Vector{Int64}}

Return the observed variables and values in uai_evid_filepath. If the passed file path is an empty string, return empty vectors.

The UAI file formats are defined in: https://personal.utdallas.edu/~vibhav.gogate/uai16-evaluation/uaiformat.html

TensorInference.read_uai_file — Method

read_uai_file(
    uai_filepath
) -> Tuple{Int64, Vector{Int64}, Int64, Any}

Parse the problem instance found in uai_filepath defined in the UAI model format.

The UAI file formats are defined in: https://personal.utdallas.edu/~vibhav.gogate/uai16-evaluation/uaiformat.html

TensorInference.read_uai_mar_file — Method

read_uai_mar_file(
    uai_mar_filepath::AbstractString
) -> Vector{Vector{Float64}}

Return the marginals of all variables. The order of the variables is the same as in the model

The UAI file formats are defined in: https://personal.utdallas.edu/~vibhav.gogate/uai16-evaluation/uaiformat.html

TensorInference.read_uai_problem — Method

read_uai_problem(
    problem::AbstractString
) -> TensorInference.UAIInstance{Float64, _A} where _A<:(TensorInference.Factor{Float64})

Read a UAI problem from an artifact.

TensorInference.rescale_array — Method

rescale_array(tensor::AbstractArray{T}) -> RescaledArray

Returns a rescaled array that equivalent to the input tensor.

TensorInference.set_evidence! — Method

set_evidence!(
    uai::TensorInference.UAIInstance,
    pairs::Pair{Int64}...
) -> TensorInference.UAIInstance

Set the evidence of an UAI instance.

Examples

julia> problem = read_uai_problem("Promedus_14"); problem.obsvars, problem.obsvals
([42, 48, 27, 30, 29, 15, 124, 5, 148], [1, 1, 1, 1, 1, 1, 1, 1, 1])

julia> set_evidence!(problem, 2=>0, 4=>1); problem.obsvars, problem.obsvals
([2, 4], [0, 1])

TensorInference.uai_problem_from_file — Method

uai_problem_from_file(
    uai_filepath::String
) -> TensorInference.UAIInstance{Float64, _A} where _A<:(TensorInference.Factor{Float64})

Read a UAI problem from a file.