Internal API

Index

Types

TensorInference.Factor

TensorInference.Samples

Functions

TensorInference.backward_sampling!

TensorInference.backward_tropical

TensorInference.parse_mar_solution_file

TensorInference.read_query_file

TensorInference.rescale_array

Types

TensorInference.Factor — Type

struct Factor{T, N}

Fields

vars

vals

Encodes a discrete function over the set of variables vars that maps each instantiation of vars into a nonnegative number in vals.

TensorInference.Samples — Type

struct Samples{L}

Fields

samples::Matrix{Int64}

labels::Vector

setmask::BitVector

The sampled configurations are stored in samples, which is a vector of vector. labels is a vector of variable names for labeling configurations. The setmask is an boolean indicator to denote whether the sampling process of a variable is complete.

Functions

TensorInference.backward_sampling! — Method

backward_sampling!(
    ixs,
    xs::Tuple,
    iy,
    y,
    samples::TensorInference.Samples,
    size_dict
) -> TensorInference.Samples

The backward process for sampling configurations.

ixs and xs are labels and tensor data for input tensors,

iy and y are labels and tensor data for the output tensor,

samples is the samples generated for eliminated variables,

size_dict is a key-value map from tensor label to dimension size.

TensorInference.backward_tropical — Method

backward_tropical(
    ixs,
    xs::Tuple,
    iy,
    y,
    ymask,
    size_dict
) -> Vector{Any}

The backward rule for tropical einsum.

ixs and xs are labels and tensor data for input tensors,

iy and y are labels and tensor data for the output tensor,

ymask is the boolean mask for gradients,

size_dict is a key-value map from tensor label to dimension size.

TensorInference.parse_mar_solution_file — Method

parse_mar_solution_file(
    rawlines::Vector{String};
    factor_eltype
) -> Vector{Vector{Float64}}

Parse the solution marginals of all variables from the UAI MAR solution file. The order of the variables is the same as in the model definition.

The UAI file formats are defined in: https://uaicompetition.github.io/uci-2022/file-formats/

TensorInference.read_query_file — Method

read_query_file(
    query_filepath::AbstractString
) -> Vector{Int64}

Return the query variables in query_filepath. If the passed file path is an empty string, return an empty vector.

The UAI file formats are defined in: https://uaicompetition.github.io/uci-2022/file-formats/

TensorInference.rescale_array — Method

rescale_array(tensor::AbstractArray{T}) -> RescaledArray

Returns a rescaled array that equivalent to the input tensor.