Public API

Index

Modules

TensorInference

Types

OMEinsumContractionOrders.GreedyMethod

OMEinsumContractionOrders.KaHyParBipartite

OMEinsumContractionOrders.MergeGreedy

OMEinsumContractionOrders.MergeVectors

OMEinsumContractionOrders.SABipartite

OMEinsumContractionOrders.TreeSA

TensorInference.MMAPModel

TensorInference.RescaledArray

TensorInference.TensorNetworkModel

TensorInference.UAIInstance

Functions

OMEinsumContractionOrders.contraction_complexity

TensorInference.get_cards

TensorInference.get_vars

TensorInference.log_probability

TensorInference.marginals

TensorInference.maximum_logp

TensorInference.most_probable_config

TensorInference.probability

TensorInference.read_evidence_file

TensorInference.read_instance

TensorInference.read_model_file

TensorInference.read_solution_file

TensorInference.read_td_file

TensorInference.sample

TensorInference.set_evidence!

Modules

TensorInference — Module

Main module for TensorInference.jl – A toolbox for probabilistic inference using contraction of tensor networks.

Exports

GreedyMethod

KaHyParBipartite

MMAPModel

MergeGreedy

MergeVectors

RescaledArray

SABipartite

TensorNetworkModel

TreeSA

UAIInstance

contraction_complexity

get_cards

get_vars

log_probability

marginals

maximum_logp

most_probable_config

probability

read_evidence_file

read_instance

read_model_file

read_solution_file

read_td_file

sample

set_evidence!

Types

OMEinsumContractionOrders.GreedyMethod — Type

GreedyMethod{MT}
GreedyMethod(; method=MinSpaceOut(), nrepeat=10)

The fast but poor greedy optimizer. Input arguments are

method is MinSpaceDiff() or MinSpaceOut.

MinSpaceOut choose one of the contraction that produces a minimum output tensor size,

MinSpaceDiff choose one of the contraction that decrease the space most.

nrepeat is the number of repeatition, returns the best contraction order.

OMEinsumContractionOrders.KaHyParBipartite — Type

KaHyParBipartite{RT,IT,GM}
KaHyParBipartite(; sc_target, imbalances=collect(0.0:0.005:0.8),
    max_group_size=40, greedy_config=GreedyMethod())

Optimize the einsum code contraction order using the KaHyPar + Greedy approach. This program first recursively cuts the tensors into several groups using KaHyPar, with maximum group size specifed by max_group_size and maximum space complexity specified by sc_target, Then finds the contraction order inside each group with the greedy search algorithm. Other arguments are

sc_target is the target space complexity, defined as log2(number of elements in the largest tensor),

imbalances is a KaHyPar parameter that controls the group sizes in hierarchical bipartition,

max_group_size is the maximum size that allowed to used greedy search,

greedy_config is a greedy optimizer.

References

Hyper-optimized tensor network contraction

Simulating the Sycamore quantum supremacy circuits

OMEinsumContractionOrders.MergeGreedy — Type

MergeGreedy <: CodeSimplifier
MergeGreedy(; threshhold=-1e-12)

Contraction code simplifier (in order to reduce the time of calling optimizers) that merges tensors greedily if the space complexity of merged tensors is reduced (difference smaller than the threshhold).

OMEinsumContractionOrders.MergeVectors — Type

MergeVectors <: CodeSimplifier
MergeVectors()

Contraction code simplifier (in order to reduce the time of calling optimizers) that merges vectors to closest tensors.

OMEinsumContractionOrders.SABipartite — Type

SABipartite{RT,BT}
SABipartite(; sc_target=25, ntrials=50, βs=0.1:0.2:15.0, niters=1000
    max_group_size=40, greedy_config=GreedyMethod(), initializer=:random)

Optimize the einsum code contraction order using the Simulated Annealing bipartition + Greedy approach. This program first recursively cuts the tensors into several groups using simulated annealing, with maximum group size specifed by max_group_size and maximum space complexity specified by sc_target, Then finds the contraction order inside each group with the greedy search algorithm. Other arguments are

size_dict, a dictionary that specifies leg dimensions,

sc_target is the target space complexity, defined as log2(number of elements in the largest tensor),

max_group_size is the maximum size that allowed to used greedy search,

βs is a list of inverse temperature 1/T,

niters is the number of iteration in each temperature,

ntrials is the number of repetition (with different random seeds),

greedy_config configures the greedy method,

initializer, the partition configuration initializer, one can choose :random or :greedy (slow but better).

References

Hyper-optimized tensor network contraction

OMEinsumContractionOrders.TreeSA — Type

TreeSA{RT,IT,GM}
TreeSA(; sc_target=20, βs=collect(0.01:0.05:15), ntrials=10, niters=50,
    sc_weight=1.0, rw_weight=0.2, initializer=:greedy, greedy_config=GreedyMethod(; nrepeat=1))

Optimize the einsum contraction pattern using the simulated annealing on tensor expression tree.

sc_target is the target space complexity,

ntrials, βs and niters are annealing parameters, doing ntrials indepedent annealings, each has inverse tempteratures specified by βs, in each temperature, do niters updates of the tree.

sc_weight is the relative importance factor of space complexity in the loss compared with the time complexity.

rw_weight is the relative importance factor of memory read and write in the loss compared with the time complexity.

initializer specifies how to determine the initial configuration, it can be :greedy or :random. If it is using :greedy method to generate the initial configuration, it also uses two extra arguments greedy_method and greedy_nrepeat.

nslices is the number of sliced legs, default is 0.

fixed_slices is a vector of sliced legs, default is [].

References

Recursive Multi-Tensor Contraction for XEB Verification of Quantum Circuits

TensorInference.MMAPModel — Type

struct MMAPModel{LT, AT<:AbstractArray}

Computing the most likely assignment to the query variables, Xₘ ⊆ X after marginalizing out the remaining variables Xₛ = X \ Xₘ.

\[{\rm MMAP}(X_i|E=e) = \arg \max_{X_M} \sum_{X_S} \prod_{F} f(x_M, x_S, e)\]

Fields

vars is the query variables in the tensor network.

code is the tropical tensor network contraction pattern.

tensors is the tensors fed into the tensor network.

clusters is the clusters, each element of this cluster is a TensorNetworkModel instance for marginalizing certain variables.

fixedvertices is a dictionary to specifiy degree of freedoms fixed to certain values, which should not have overlap with the query variables.

TensorInference.RescaledArray — Type

struct RescaledArray{T, N, AT<:AbstractArray{T, N}} <: AbstractArray{T, N}

RescaledArray(α, T) -> RescaledArray

An array data type with a log-prefactor, and a l∞-normalized storage, i.e. the maximum element in a tensor is 1. This tensor type can avoid the potential underflow/overflow of numbers in a tensor network. The constructor RescaledArray(α, T) creates a rescaled array that equal to exp(α) * T.

TensorInference.TensorNetworkModel — Type

struct TensorNetworkModel{LT, ET, MT<:AbstractArray}

Probabilistic modeling with a tensor network.

Fields

vars is the degree of freedoms in the tensor network.

code is the tensor network contraction pattern.

tensors is the tensors fed into the tensor network.

fixedvertices is a dictionary to specifiy degree of freedoms fixed to certain values.

TensorInference.UAIInstance — Type

struct UAIInstance{ET, FT<:(TensorInference.Factor{ET})}

Fields

nvars is the number of variables,

nclique is the number of cliques,

cards is a vector of cardinalities for variables,

factors is a vector of factors,

obsvars is a vector of observed variables,

obsvals is a vector of observed values,

queryvars is a vector of query variables,

reference_solution is a vector with the reference solution.

Functions

OMEinsumContractionOrders.contraction_complexity — Function

contraction_complexity(tensor_network)

Returns the contraction complexity of a tensor newtork model.

contraction_complexity(eincode, size_dict) -> ContractionComplexity

Returns the time, space and read-write complexity of the einsum contraction. The returned object contains 3 fields:

time complexity tc defined as log2(number of element-wise multiplications).

space complexity sc defined as log2(size of the maximum intermediate tensor).

read-write complexity rwc defined as log2(the number of read-write operations).

TensorInference.get_cards — Function

get_cards(tn::TensorNetworkModel; fixedisone) -> Vector

Get the cardinalities of variables in this tensor network.

get_cards(mmap::MMAPModel; fixedisone) -> Vector

TensorInference.get_vars — Function

get_vars(tn::TensorNetworkModel) -> Vector

Get the variables in this tensor network, they are also known as legs, labels, or degree of freedoms.

get_vars(mmap::MMAPModel) -> Vector

TensorInference.log_probability — Function

log_probability(
    tn::TensorNetworkModel,
    config::Union{Dict, AbstractVector}
) -> Real

Evaluate the log probability of config.

TensorInference.marginals — Function

marginals(
    tn::TensorNetworkModel;
    usecuda,
    rescale
) -> Vector

Returns the marginal probability distribution of variables. One can use get_vars(tn) to get the full list of variables in this tensor network.

TensorInference.maximum_logp — Function

maximum_logp(
    tn::TensorNetworkModel;
    usecuda
) -> AbstractArray{<:Real}

Returns an output array containing largest log-probabilities.

TensorInference.most_probable_config — Function

most_probable_config(
    tn::TensorNetworkModel;
    usecuda
) -> Tuple{Real, Vector}

Returns the largest log-probability and the most probable configuration.

TensorInference.probability — Function

probability(
    tn::TensorNetworkModel;
    usecuda,
    rescale
) -> AbstractArray

Contract the tensor network and return a probability array with its rank specified in the contraction code tn.code. The returned array may not be l1-normalized even if the total probability is l1-normalized, because the evidence tn.fixedvertices may not be empty.

TensorInference.read_evidence_file — Function

read_evidence_file(
    evidence_filepath::AbstractString
) -> Tuple{Vector{Int64}, Vector{Int64}}

Return the observed variables and values in evidence_filepath. If the passed file path is an empty string, return empty vectors.

The UAI file formats are defined in: https://uaicompetition.github.io/uci-2022/file-formats/

TensorInference.read_instance — Function

read_instance(
    model_filepath::AbstractString;
    evidence_filepath,
    query_filepath,
    solution_filepath,
    eltype
) -> UAIInstance

Read a UAI problem instance from a file.

TensorInference.read_model_file — Function

read_model_file(
    model_filepath;
    factor_eltype
) -> Tuple{Int64, Vector{Int64}, Int64, Any}

Parse the problem instance found in model_filepath defined in the UAI model format.

The UAI file formats are defined in: https://uaicompetition.github.io/uci-2022/file-formats/

TensorInference.read_solution_file — Function

read_solution_file(
    solution_filepath::AbstractString;
    factor_eltype
) -> Union{Float64, Vector{Vector{Float64}}, Vector{Int64}}

Return the solution in solution_filepath. Returns an empty vector if the extension is not supported.

The UAI file formats are defined in: https://uaicompetition.github.io/uci-2022/file-formats/

TensorInference.read_td_file — Function

read_td_file(
    td_filepath::AbstractString
) -> Tuple{Int64, Int64, Int64, Vector{Vector{Int64}}, Vector{Vector{Int64}}}

Parse a tree decomposition instance described the PACE format.

The PACE file format is defined in: https://pacechallenge.org/2017/treewidth/

TensorInference.sample — Function

sample(
    tn::TensorNetworkModel,
    n::Int64;
    usecuda
) -> Matrix{Int64}

Generate samples from a tensor network based probabilistic model. Returns a vector of vector, each element being a configurations defined on get_vars(tn).

Arguments

tn is the tensor network model.

n is the number of samples to be returned.

TensorInference.set_evidence! — Function

set_evidence!(
    uai::UAIInstance,
    pairs::Pair{Int64}...
) -> UAIInstance

Set the evidence of an UAI instance.